+++ src/redapid/list.h
@@ -1,6 +1,6 @@
 /*
  * redapid
- * Copyright (C) 2014 Matthias Bolte <matthias@tinkerforge.com>
+ * Copyright (C) 2014-2015 Matthias Bolte <matthias@tinkerforge.com>
  *
  * list.h: List object implementation
  *
@@ -33,7 +33,7 @@ typedef struct {
 } List;
 
 APIE list_allocate(uint16_t reserve, Session *session,
-                   uint16_t object_create_flags, ObjectID *id, List **object);
+                   uint32_t object_create_flags, ObjectID *id, List **object);
 
 APIE list_get_length(List *list, uint16_t *length);
 APIE list_get_item(List *list, uint16_t index, Session *session,
